Define interface for DAG implementations to prepare for other backends (than in-memory)
