+++ pkg/nextgen/distdoc/dag_contract_test.go
@@ -1,13 +1,12 @@
 package distdoc
 
 import (
-	"crypto"
-	"crypto/x509"
 	"encoding/binary"
 	"github.com/nuts-foundation/nuts-network/pkg/model"
 	"github.com/stretchr/testify/assert"
 	"strings"
 	"testing"
+	"time"
 )
 
 // trackingVisitor just keeps track of which nodes were visited in what order.
@@ -22,7 +21,7 @@ func (n *trackingVisitor) Accept(document Document) {
 func (n trackingVisitor) JoinRefsAsString() string {
 	var contents []string
 	for _, document := range n.documents {
-		val := strings.TrimLeft(document.Ref().String(), "0")
+		val := strings.TrimLeft(model.Hash(document.Payload()).String(), "0")
 		if val == "" {
 			val = "0"
 		}
@@ -31,31 +30,39 @@ func (n trackingVisitor) JoinRefsAsString() string {
 	return strings.Join(contents, ", ")
 }
 
-func TestChain_Add(t *testing.T) {
-	key, certificate := generateKeyAndCertificate()
+func DAGTest_Add(creator func(t *testing.T) DAG, t *testing.T) {
 	t.Run("ok", func(t *testing.T) {
-		graph := NewDAG()
+		graph := creator(t)
 		doc := testDocument(0)
 
 		err := graph.Add(doc)
 
 		assert.NoError(t, err)
 		visitor := trackingVisitor{}
-		graph.Walk(visitor.Accept)
+		err = graph.Walk(visitor.Accept)
+		if !assert.NoError(t, err) {
+			return
+		}
 		assert.Len(t, visitor.documents, 1)
 		assert.Equal(t, doc.Ref(), visitor.documents[0].Ref())
 	})
 	t.Run("ok - out of order", func(t *testing.T) {
-		_, documents := graphF(key, certificate)
-		graph := NewDAG()
+		_, documents := graphF(creator, t)
+		graph := creator(t)
 
 		for i := len(documents) - 1; i >= 0; i-- {
-			graph.Add(documents[i])
+			err := graph.Add(documents[i])
+			if !assert.NoError(t, err) {
+				return
+			}
 		}
 
 		visitor := trackingVisitor{}
-		graph.Walk(visitor.Accept)
-		assert.Equal(t, "0, 1, 2, 3, 4, 5", visitor.JoinRefsAsString())
+		err := graph.Walk(visitor.Accept)
+		if !assert.NoError(t, err) {
+			return
+		}
+		assert.Regexp(t, "0, (1, 2|2, 1), (3, 4|4, 3), 5", visitor.JoinRefsAsString())
 	})
 	t.Run("error - cyclic graph", func(t *testing.T) {
 		t.Skip("Algorithm for detecting cycles is not yet decided on")
@@ -65,22 +72,23 @@ func TestChain_Add(t *testing.T) {
 		C := testDocument(2, B.Ref())
 		B.prevs = append(B.prevs, C.Ref())
 
-		graph := NewDAG()
+		graph := creator(t)
 		err := graph.Add(A, B, C)
 		assert.EqualError(t, err, "")
 	})
 }
 
-func TestChain_Walk(t *testing.T) {
-	key, certificate := generateKeyAndCertificate()
-
+func DAGTest_Walk(creator func(t *testing.T) DAG, t *testing.T) {
 	t.Run("ok - walk graph F", func(t *testing.T) {
 		visitor := trackingVisitor{}
-		graph, _ := graphF(key, certificate)
+		graph, _ := graphF(creator, t)
 
-		graph.Walk(visitor.Accept)
+		err := graph.Walk(visitor.Accept)
+		if !assert.NoError(t, err) {
+			return
+		}
 
-		assert.Equal(t, "0, 1, 2, 3, 4, 5", visitor.JoinRefsAsString())
+		assert.Regexp(t, "0, (1, 2|2, 1), (3, 4|4, 3), 5", visitor.JoinRefsAsString())
 	})
 
 	t.Run("ok - walk graph G", func(t *testing.T) {
@@ -94,13 +102,13 @@ func TestChain_Walk(t *testing.T) {
 		//...................\.../
 		//.....................G
 		visitor := trackingVisitor{}
-		graph, docs := graphF(key, certificate)
+		graph, docs := graphF(creator, t)
 		G := testDocument(6, docs[3].Ref(), docs[5].Ref())
 		graph.Add(G)
 
 		graph.Walk(visitor.Accept)
 
-		assert.Equal(t, "0, 1, 2, 3, 4, 5, 6", visitor.JoinRefsAsString())
+		assert.Regexp(t, "0, (1, 2|2, 1), (3, 4|4, 3), 5, 6", visitor.JoinRefsAsString())
 	})
 
 	t.Run("ok - walk graph F, C is missing", func(t *testing.T) {
@@ -112,8 +120,8 @@ func TestChain_Walk(t *testing.T) {
 		//.......................\
 		//........................F
 		visitor := trackingVisitor{}
-		_, docs := graphF(key, certificate)
-		graph := NewDAG()
+		_, docs := graphF(creator, t)
+		graph := creator(t)
 		graph.Add(docs[0], docs[1], docs[3], docs[4], docs[5])
 
 		graph.Walk(visitor.Accept)
@@ -122,20 +130,22 @@ func TestChain_Walk(t *testing.T) {
 	})
 
 	t.Run("ok - empty graph", func(t *testing.T) {
-		graph := NewDAG()
+		graph := creator(t)
 		visitor := trackingVisitor{}
 
-		graph.Walk(visitor.Accept)
+		err := graph.Walk(visitor.Accept)
+		if !assert.NoError(t, err) {
+			return
+		}
 
 		assert.Empty(t, visitor.documents)
 	})
 
 	t.Run("ok - document added twice", func(t *testing.T) {
-		graph := NewDAG()
+		graph := creator(t)
 		d := testDocument(0)
 		graph.Add(d)
-		err := graph.Add(d)
-		assert.NoError(t, err)
+		graph.Add(d)
 		visitor := trackingVisitor{}
 
 		graph.Walk(visitor.Accept)
@@ -144,11 +154,12 @@ func TestChain_Walk(t *testing.T) {
 	})
 
 	t.Run("error - second root document", func(t *testing.T) {
-		graph := NewDAG()
-		d1 := testDocument(0 )
+		graph := creator(t)
+		d1 := testDocument(0)
 		d2 := testDocument(1)
-		graph.Add(d1)
-		err := graph.Add(d2)
+		err := graph.Add(d1)
+
+		err = graph.Add(d2)
 		assert.Equal(t, errRootAlreadyExists, err)
 		visitor := trackingVisitor{}
 
@@ -159,22 +170,21 @@ func TestChain_Walk(t *testing.T) {
 	})
 }
 
-
-func TestDAG_MissingDocuments(t *testing.T) {
+func DAGTest_MissingDocuments(creator func(t *testing.T) DAG, t *testing.T) {
 	A := testDocument(0)
 	B := testDocument(1, A.Ref())
 	C := testDocument(2, B.Ref())
 	t.Run("no missing documents (empty graph)", func(t *testing.T) {
-		graph := NewDAG()
+		graph := creator(t)
 		assert.Empty(t, graph.MissingDocuments())
 	})
 	t.Run("no missing documents (non-empty graph)", func(t *testing.T) {
-		graph := NewDAG()
+		graph := creator(t)
 		graph.Add(A, B, C)
 		assert.Empty(t, graph.MissingDocuments())
 	})
 	t.Run("missing documents (non-empty graph)", func(t *testing.T) {
-		graph := NewDAG()
+		graph := creator(t)
 		graph.Add(A, C)
 		assert.Len(t, graph.MissingDocuments(), 1)
 		// Now add missing document B and assert there are no more missing documents
@@ -191,8 +201,8 @@ func TestDAG_MissingDocuments(t *testing.T) {
 //.................D    E
 //.......................\
 //........................F
-func graphF(key crypto.Signer, certificate *x509.Certificate) (*DAG, []Document) {
-	graph := NewDAG()
+func graphF(creator func(t *testing.T) DAG, t *testing.T) (DAG, []Document) {
+	graph := creator(t)
 	A := testDocument(0)
 	B := testDocument(1, A.Ref())
 	C := testDocument(2, A.Ref())
@@ -204,13 +214,12 @@ func graphF(key crypto.Signer, certificate *x509.Certificate) (*DAG, []Document)
 	return graph, docs
 }
 
+var privateKey, certificate = generateKeyAndCertificate()
+
 func testDocument(num uint32, prevs ...model.Hash) Document {
 	ref := model.EmptyHash()
 	binary.BigEndian.PutUint32(ref[model.HashSize-4:], num)
-	return &document{
-		prevs:       prevs,
-		payload:     ref,
-		payloadType: "foo/bar",
-		ref:         ref,
-	}
+	unsignedDocument, _ := NewDocument(PayloadHash(ref), "foo/bar", prevs)
+	signedDocument, _ := unsignedDocument.Sign(privateKey, certificate, time.Now())
+	return signedDocument
 }
